Create simple program to calculate employee hourly
Assignment Description:
Develop a program that will determine the gross pay for each of several
employees. The company pays “straight time” for the first 40 hours
worked by each employee and pays “time-and-a-half” for all hours worked
in excess of 40 hours. You’re given a list of the company’s employees,
the number of hours each worked last week and each employee’s hourly
rate. Your program should use scanf to input this information for each
employee and determine and display the employee’s gross pay.